requirements.txt
nvitop
PyYaml
chardet
loguru

setuptools
wheel
